+++ integration/fieldoption-jstype-with-forcelong-bigint/fieldoption-jstype-with-forcelong-bigint.ts
@@ -51,21 +51,21 @@ export const FieldOption = {
             break;
           }
 
-          message.normalField = longToBigint(reader.int64() as Long);
+          message.normalField = longToBigint(reader.int64());
           continue;
         case 2:
           if (tag !== 16) {
             break;
           }
 
-          message.numberField = longToNumber(reader.int64() as Long);
+          message.numberField = longToNumber(reader.int64());
           continue;
         case 3:
           if (tag !== 24) {
             break;
           }
 
-          message.stringField = longToString(reader.int64() as Long);
+          message.stringField = longToString(reader.int64());
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
