+++ integration/simple-long-string/google/protobuf/timestamp.ts
@@ -140,7 +140,7 @@ export const Timestamp = {
             break;
           }
 
-          message.seconds = longToString(reader.int64() as Long);
+          message.seconds = longToString(reader.int64());
           continue;
         case 2:
           if (tag !== 16) {
