+++ integration/simple-long-string/google/protobuf/wrappers.ts
@@ -234,7 +234,7 @@ export const Int64Value = {
             break;
           }
 
-          message.value = longToString(reader.int64() as Long);
+          message.value = longToString(reader.int64());
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
@@ -291,7 +291,7 @@ export const UInt64Value = {
             break;
           }
 
-          message.value = longToString(reader.uint64() as Long);
+          message.value = longToString(reader.uint64());
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
