+++ integration/simple-snake/simple.ts
@@ -1779,7 +1779,7 @@ export const Numbers = {
             break;
           }
 
-          message.int64 = longToNumber(reader.int64() as Long);
+          message.int64 = longToNumber(reader.int64());
           continue;
         case 5:
           if (tag !== 40) {
@@ -1793,7 +1793,7 @@ export const Numbers = {
             break;
           }
 
-          message.uint64 = longToNumber(reader.uint64() as Long);
+          message.uint64 = longToNumber(reader.uint64());
           continue;
         case 7:
           if (tag !== 56) {
@@ -1807,7 +1807,7 @@ export const Numbers = {
             break;
           }
 
-          message.sint64 = longToNumber(reader.sint64() as Long);
+          message.sint64 = longToNumber(reader.sint64());
           continue;
         case 9:
           if (tag !== 77) {
@@ -1821,7 +1821,7 @@ export const Numbers = {
             break;
           }
 
-          message.fixed64 = longToNumber(reader.fixed64() as Long);
+          message.fixed64 = longToNumber(reader.fixed64());
           continue;
         case 11:
           if (tag !== 93) {
@@ -1835,7 +1835,7 @@ export const Numbers = {
             break;
           }
 
-          message.sfixed64 = longToNumber(reader.sfixed64() as Long);
+          message.sfixed64 = longToNumber(reader.sfixed64());
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
