+++ integration/unknown-fields/google/protobuf/descriptor.ts
@@ -3248,14 +3248,14 @@ export const UninterpretedOption = {
             break;
           }
 
-          message.positiveIntValue = longToNumber(reader.uint64() as Long);
+          message.positiveIntValue = longToNumber(reader.uint64());
           continue;
         case 5:
           if (tag !== 40) {
             break;
           }
 
-          message.negativeIntValue = longToNumber(reader.int64() as Long);
+          message.negativeIntValue = longToNumber(reader.int64());
           continue;
         case 6:
           if (tag !== 49) {
